picolibc: Fix run-riscv script argument passing
Some tests require arguments to be passed and that wasn't working
quite right as qemu has different ideas about what the default command
line should look like, with some targets setting it to the empty
string and others setting it to the name of the loaded program.

When no arguments are provided, set it to the empty string so that
both arm and riscv get the same result in qemu.

Signed-off-by: Keith Packard <keithp@keithp.com>